Modularize nicegui app
