+++ src/gui/pages/categories.py
@@ -1,0 +1,40 @@
+from nicegui import ui
+
+from ..helpers import call_api
+from ..theme import theme
+
+
+def create() -> None:
+    @ui.page("/categories/")
+    def categories_page():
+        with theme.frame():
+            data = call_api("/categories/", method="GET")
+
+            ui.label("Categories").classes("text-xl font-bold")
+
+            with ui.column().classes("gap-2"):
+                ui.table(
+                    columns=[
+                        {
+                            "name": "id",
+                            "label": "ID",
+                            "field": "id",
+                        },
+                        {
+                            "name": "name",
+                            "label": "Name",
+                            "field": "name",
+                        },
+                        {
+                            "name": "budget",
+                            "label": "Budget",
+                            "field": "budget",
+                        },
+                    ],
+                    rows=data,
+                    row_key="id",
+                ).classes("w-full")
+
+                ui.button(
+                    "Add Category", on_click=lambda: ui.notify("TODO: Add category")
+                )
